refactor: Starknet artifact generator tests
